requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
pydantic==1.10.7
PyJWT==2.6.0
sniffio==1.3.0
SQLAlchemy==2.0.7
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
